feat(initialization): create schnack client.js
