+++ package.json
@@ -54,11 +54,11 @@
   "devDependencies": {
     "http-server": "^0.10.0",
     "nodemon": "^1.12.1",
-    "rollup": "^0.50.0",
+    "rollup": "^0.55.4",
     "rollup-plugin-buble": "^0.16.0",
-    "rollup-plugin-commonjs": "^8.2.1",
+    "rollup-plugin-commonjs": "^8.3.0",
     "rollup-plugin-jst": "^1.0.1",
-    "rollup-plugin-node-resolve": "^3.0.0",
+    "rollup-plugin-node-resolve": "^3.0.2",
     "rollup-plugin-string": "^2.0.2",
     "rollup-plugin-template": "^1.0.10",
     "rollup-plugin-uglify": "^2.0.1",
